+++ src/plone/oidc/tests/test_behaviors.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
-from plone.oidc.membrane.behaviors import IOidConnectClaims
-from plone.oidc.testing import PLONE_OIDC_INTEGRATION_TESTING
 from plone import api
 from plone.behavior.interfaces import IBehavior
 from plone.dexterity.behavior import DexterityBehaviorAssignable
+from plone.oidc.membrane.behaviors import IOidConnectClaims
+from plone.oidc.testing import PLONE_OIDC_INTEGRATION_TESTING
 from zope.component import queryUtility
-from zope.interface import Interface
 
 import unittest
 
@@ -31,55 +30,26 @@ class TestBehaviors(unittest.TestCase):
         oid_connection_claims = queryUtility(IBehavior, name='plone.oidc.membrane.behaviors.IOidConnectClaims')
         self.assertIsNotNone(oid_connection_claims)
 
-
-    def offtest_supports(self):
+    def test_supports(self):
         """ """
         # Context mock
         with api.env.adopt_roles('Manager'):
-            id_ = self.portal.invokeFactory('Organization', 'hospital')
-            organization_context = self.portal[id_]
-            id_ = organization_context.invokeFactory('Patient', 'patient')
-            patient_context = organization_context[id_]
+            id_ = self.portal.invokeFactory('OidcTestUser', 'test_oidc_user')
+            user_context = self.portal[id_]
 
-        # Test: Organization has IHeartIdChooser but not IOidConnectClaims
-        assignable = DexterityBehaviorAssignable(organization_context)
-        self.assertEqual(False, assignable.supports(IOidConnectClaims))
-        self.assertEqual(True, assignable.supports(IHeartIdChooser))
-
-        # Test: Patient has both behaviors
-        assignable = DexterityBehaviorAssignable(patient_context)
+        # Test: OidcTestUser has expected behaviors
+        assignable = DexterityBehaviorAssignable(user_context)
         self.assertEqual(True, assignable.supports(IOidConnectClaims))
-        self.assertEqual(True, assignable.supports(IHeartIdChooser))
 
-    def offtest_enumerate(self):
+    def test_enumerate(self):
         """ """
         with api.env.adopt_roles('Manager'):
-            id_ = self.portal.invokeFactory('Organization', 'hospital')
-            organization_context = self.portal[id_]
-            id_ = organization_context.invokeFactory('Patient', 'patient')
-            patient_context = organization_context[id_]
+            id_ = self.portal.invokeFactory('OidcTestUser', 'test_oidc_user2')
+            user_context = self.portal[id_]
 
         oid_connection_claims = queryUtility(IBehavior, name=IOidConnectClaims.__identifier__)
-        id_chooser = queryUtility(IBehavior, name=IHeartIdChooser.__identifier__)
 
-        assignable = DexterityBehaviorAssignable(organization_context)
-
-        # Organization has IHeartIdChooser
-        self.assertIn(
-            id_chooser,
-            list(assignable.enumerateBehaviors())
-        )
-        self.assertNotIn(
-            oid_connection_claims,
-            list(assignable.enumerateBehaviors())
-        )
-
-        # Patient has both
-        assignable = DexterityBehaviorAssignable(patient_context)
-        self.assertIn(
-            id_chooser,
-            list(assignable.enumerateBehaviors())
-        )
+        assignable = DexterityBehaviorAssignable(user_context)
         self.assertIn(
             oid_connection_claims,
             list(assignable.enumerateBehaviors())