+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.0a2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- ``plone.oidc.membrane.behaviors.IOidConnectClaims`` behavior is available.
+- ``oidc_membrane_user_workflow`` workflow is available.
 
 
 1.0.0a1 (2018-07-20)